+++ src/components/ui/data-display/index.ts
@@ -8,3 +8,4 @@ export * from './Pagination';
 export * from './Progress';
 export * from './Skeleton';
 export * from './GoogleMap';
+export * from './ProofTile';